agregando eureka y gateway

+++ allmusic/src/main/java/ar/edu/unnoba/poo2024/allmusic/resource/SongResource.java
@@ -1,16 +1,18 @@
 package ar.edu.unnoba.poo2024.allmusic.resource;
 
 import ar.edu.unnoba.poo2024.allmusic.dto.SongResponseDTO;
+import ar.edu.unnoba.poo2024.allmusic.model.MusicArtistUser;
 import ar.edu.unnoba.poo2024.allmusic.dto.SongRequestDTO;
 import ar.edu.unnoba.poo2024.allmusic.service.SongService;
+import ar.edu.unnoba.poo2024.allmusic.service.UserService;
+import ar.edu.unnoba.poo2024.allmusic.util.JwtTokenUtil;
+import ar.edu.unnoba.poo2024.allmusic.util.ResourceNotFoundException;
 import ar.edu.unnoba.poo2024.allmusic.service.AuthorizationService;
-import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api/song")
@@ -23,29 +25,35 @@ public class SongResource {
     private AuthorizationService authorizationService;
 
     @Autowired
-    private ModelMapper modelMapper;
+    private UserService userService;
+
+    @Autowired
+    private JwtTokenUtil jwtTokenUtil;
 
     @GetMapping("/songs")
-    public ResponseEntity<?> getSongs(@RequestHeader("Authorization") String tokenJWT) {
+    public ResponseEntity<?> getSongs(@RequestHeader("Authorization") String tokenJWT,
+            @RequestParam(value = "artist", required = false) String artistName,
+            @RequestParam(value = "genre", required = false) String genre) {
         try {
-            // Eliminar el prefijo Bearer
             String token = tokenJWT.replace("Bearer ", "");
-
-            // Validar el token JWT
+            String userType = jwtTokenUtil.getUserType(token);
+            if (!userType.equals("Artist")) {
+                return new ResponseEntity<>("Acces forbidden for Artist", HttpStatus.FORBIDDEN);
+            }
             authorizationService.verify(token);
+            List<SongResponseDTO> songResponseDTO;
+            if (artistName != null && genre != null) {
+                songResponseDTO = songService.getSongsByArtistAndGenre(artistName, genre);
+            } else if (artistName != null) {
+                songResponseDTO = songService.getSongsByArtist(artistName);
+            } else if (genre != null) {
+                songResponseDTO = songService.getSongsByGenre(genre);
+            } else {
+                songResponseDTO = songService.getAllSongs();
+            }
+            return new ResponseEntity<>(songResponseDTO, HttpStatus.OK);
 
-            // Obtener todas las instancias de Song
-            List<SongResponseDTO> songs = songService.getAllSongs();
-
-            // Mapear la lista de Song a SongResponseDTO
-            List<SongResponseDTO> songResponseDTOs = songs.stream()
-                    .map(song -> modelMapper.map(song, SongResponseDTO.class))
-                    .collect(Collectors.toList());
-
-            // Retornar la lista de SongResponseDTO con código de estado 200
-            return new ResponseEntity<>(songResponseDTOs, HttpStatus.OK);
         } catch (Exception e) {
-            // Retornar código de estado 403 si la autorización falla
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
     }
@@ -53,64 +61,52 @@ public class SongResource {
     @GetMapping("/{id}")
     public ResponseEntity<?> getSongById(@PathVariable Long id, @RequestHeader("Authorization") String tokenJWT) {
         try {
-            // Eliminar el prefijo Bearer
             String token = tokenJWT.replace("Bearer ", "");
-
-            // Validar el token JWT
             authorizationService.verify(token);
-
-            // Obtener la instancia de Song por id
-            SongResponseDTO song = songService.getSongById(id).orElseThrow(() -> new RuntimeException("Song not found"));
-
-            // Mapear la instancia de Song a SongResponseDTO
-            SongResponseDTO songResponseDTO = modelMapper.map(song, SongResponseDTO.class);
-
-            // Retornar la instancia de SongResponseDTO con código de estado 200
+            SongResponseDTO songResponseDTO = songService.getSongById(id)
+                    .orElseThrow(() -> new RuntimeException("Song not found"));
             return new ResponseEntity<>(songResponseDTO, HttpStatus.OK);
         } catch (Exception e) {
-            // Retornar código de estado 403 si la autorización falla
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
     }
 
     @PostMapping
-    public ResponseEntity<?> createSong(@RequestBody SongResponseDTO songResponseDTO, @RequestHeader("Authorization") String tokenJWT) {
+    public ResponseEntity<?> createSong(@RequestBody SongRequestDTO songRequestDTO,
+            @RequestHeader("Authorization") String tokenJWT) {
         try {
-            // Eliminar el prefijo Bearer
             String token = tokenJWT.replace("Bearer ", "");
-
-            // Validar el token JWT
+            String userType = jwtTokenUtil.getUserType(token);
+            if (!userType.equals("Artist")) {
+                System.out.println("Access forbidden for non-Artist users");
+                return new ResponseEntity<>("Access forbidden for non-Artist users", HttpStatus.FORBIDDEN);
+            }
             authorizationService.verify(token);
-
-            // Crear una nueva instancia de Song
-            SongRequestDTO songRequestDTO = modelMapper.map(songResponseDTO, SongRequestDTO.class);
             SongResponseDTO createdSong = songService.createSong(songRequestDTO);
-
-            // Retornar la instancia de SongResponseDTO con código de estado 201
             return new ResponseEntity<>(createdSong, HttpStatus.CREATED);
+        } catch (ResourceNotFoundException e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
+        } catch (IllegalArgumentException e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
         } catch (Exception e) {
-            // Retornar código de estado 403 si la autorización falla
-            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<?> updateSong(@PathVariable Long id, @RequestBody SongResponseDTO songResponseDTO, @RequestHeader("Authorization") String tokenJWT) {
+    public ResponseEntity<?> updateSong(@PathVariable Long id, @RequestBody SongRequestDTO songRequestDTO,
+            @RequestHeader("Authorization") String tokenJWT) {
         try {
-            // Eliminar el prefijo Bearer
             String token = tokenJWT.replace("Bearer ", "");
-
-            // Validar el token JWT
+            String userType = jwtTokenUtil.getUserType(token);
+            if (!userType.equals("Artist")) {
+                System.out.println("Access forbidden for non-Artist users");
+                return new ResponseEntity<>("Access forbidden for non-Artist users", HttpStatus.FORBIDDEN);
+            }
             authorizationService.verify(token);
-
-            // Actualizar la instancia de Song por id
-            SongRequestDTO songRequestDTO = modelMapper.map(songResponseDTO, SongRequestDTO.class);
             SongResponseDTO updatedSong = songService.updateSong(id, songRequestDTO);
-
-            // Retornar la instancia de SongResponseDTO con código de estado 200
             return new ResponseEntity<>(updatedSong, HttpStatus.OK);
         } catch (Exception e) {
-            // Retornar código de estado 403 si la autorización falla
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
     }
@@ -118,19 +114,17 @@ public class SongResource {
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteSong(@PathVariable Long id, @RequestHeader("Authorization") String tokenJWT) {
         try {
-            // Eliminar el prefijo Bearer
             String token = tokenJWT.replace("Bearer ", "");
+            String userType = jwtTokenUtil.getUserType(token);
 
-            // Validar el token JWT
+            if (!userType.equals("Artist")) {
+                System.out.println("Access forbidden for non-Artist users");
+                return new ResponseEntity<>("Access forbidden for non-Artist users", HttpStatus.FORBIDDEN);
+            }
             authorizationService.verify(token);
-
-            // Eliminar la instancia de Song por id
             songService.deleteSong(id);
-
-            // Retornar código de estado 204
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } catch (Exception e) {
-            // Retornar código de estado 403 si la autorización falla
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
     }
@@ -138,26 +132,19 @@ public class SongResource {
     @GetMapping("/me-songs")
     public ResponseEntity<?> getMySongs(@RequestHeader("Authorization") String tokenJWT) {
         try {
-            // Eliminar el prefijo Bearer
             String token = tokenJWT.replace("Bearer ", "");
-
-            // Validar el token JWT
+            String userType = jwtTokenUtil.getUserType(token);
+            if (!userType.equals("Artist")) {
+                return new ResponseEntity<>("Acces forbidden for User", HttpStatus.FORBIDDEN);
+            }
             authorizationService.verify(token);
+            Long artistId = Long.parseLong(jwtTokenUtil.getSubject(token));
+            MusicArtistUser artist = (MusicArtistUser) userService.findById(artistId);
 
-            // Obtener todas las instancias de Song
-            List<SongResponseDTO> songs = songService.getAllSongs();
-
-            // Mapear la lista de Song a SongResponseDTO
-            List<SongResponseDTO> songResponseDTOs = songs.stream()
-                    .map(song -> modelMapper.map(song, SongResponseDTO.class))
-                    .collect(Collectors.toList());
-
-            // Retornar la lista de SongResponseDTO con código de estado 200
+            List<SongResponseDTO> songResponseDTOs = songService.getSongsByArtist(artist.getUsername());
             return new ResponseEntity<>(songResponseDTOs, HttpStatus.OK);
         } catch (Exception e) {
-            // Retornar código de estado 403 si la autorización falla
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
     }
-
 }
